Organization table

+++ src/Layouts/Table/index.js
@@ -1,18 +1,30 @@
 import Table from "../../Components/Table";
 import { useState } from "react";
 
-export default function TableLayout({ onDeleteToast, onCreateToast }) {
+export default function TableLayout({
+  onDeleteToast,
+  onCreateToast,
+  getObjects,
+  onDeleteObject,
+  onCreateObject,
+  formSubmitText,
+  createObjectFormFields,
+  createObjectValidationSchema,
+  addObjectText,
+  tableHeader,
+  searchPlaceholder,
+}) {
   const [searchValue, setSearch] = useState("");
   const [sortValue, setSort] = useState(true);
 
-  const onDeleteObject = (object) => {
-    console.log(`${object.fullname} - deleted`);
-    onDeleteToast();
+  const onDelete = (object) => {
+    onDeleteObject(object);
+    if (onDeleteToast) onDeleteToast();
   };
 
-  const onCreateObject = (values) => {
-    console.log(values);
-    onCreateToast();
+  const onCreate = (values) => {
+    if (onCreateToast) onCreateToast();
+    return onCreateObject(values);
   };
 
   return (
@@ -20,11 +32,18 @@ export default function TableLayout({ onDeleteToast, onCreateToast }) {
       searchValue={searchValue}
       onSearchChange={(event) => setSearch(event.target.value)}
       sortValue={sortValue}
-      sortField={"fullname"}
-      searchField={"fullname"}
+      sortField={"lastName"}
+      searchField={"lastName"}
       onSortChange={() => setSort((prev) => !prev)}
-      onDeleteObject={onDeleteObject}
-      onCreateObject={onCreateObject}
+      onDeleteObject={onDelete}
+      onCreateObject={onCreate}
+      getObjects={getObjects}
+      formSubmitText={formSubmitText}
+      createObjectFormFields={createObjectFormFields}
+      createObjectValidationSchema={createObjectValidationSchema}
+      addObjectText={addObjectText}
+      tableHeader={tableHeader}
+      searchPlaceholder={searchPlaceholder}
     />
   );
 }
